Move more functions to association helper
