+++ open_alchemy/schemas/validation/association.py
@@ -1,0 +1,1 @@
+"""Validate association tables."""
