+++ tests/open_alchemy/schemas/helpers/association/test_calculate_property_schema.py
@@ -1,4 +1,4 @@
-"""Tests for the association pre-processor."""
+"""Tests for the calculate_property_schema association helper."""
 
 import pytest
